Flatten the hierarchy
My strategy for building the website relies crucially on _output.yaml specifying common options and putting shared dependencies in libs. Putting [Rr]md files down in directories doesn't play nice with that. Current workaround is to put anything that needs to be rendered in top-level directory. Not sure this is sustainable.